Patch 3: Move Project Global Client Code into it's own module
