+++ android/gradle.properties
@@ -16,7 +16,8 @@ org.gradle.jvmargs=-Xmx16g
 # https://developer.android.com/topic/libraries/support-library/androidx-rn
 android.useAndroidX=true
 # Automatically convert third-party libraries to use AndroidX
-android.enableJetifier=false
+android.enableJetifier=true
+android.jetifier.blacklist=bcprov
 # Kotlin code style for this project: "official" or "obsolete":
 kotlin.code.style=official
 android.bundle.enableUncompressedNativeLibs=false
