+++ android/vpnClient/src/main/java/org/mozilla/firefox/vpn/qt/PackageManagerHelper.java
