+++ android/vpnglean/build.gradle
@@ -22,7 +22,7 @@ buildscript {
 }
 
 plugins {
-    id "com.jetbrains.python.envs"
+    id "com.jetbrains.python.envs" version "0.0.26"
 }
 
 // Copied from from 3rdparty/glean/build.gradle
